@@ -1,26 +1,26 @@
 // Copyright © Fleuronic LLC. All rights reserved.
 
+import AppKit
+
 import struct Raindrop.Raindrop
 import struct Raindrop.Collection
 import struct Foundation.Selector
 import class Foundation.NSObject
-import class AppKit.NSMenuItem
 import struct Identity.Identifier
-import struct DewdropService.IdentifiedRaindrop
 
 public extension RaindropList {
 	protocol View: NSObject {
 		associatedtype Screen: RaindropList.Screen
 
 		var raindropAction: Selector { get }
-		var raindropItems: [Raindrop.ID: NSMenuItem] { get set }
+		var raindropItems: [Collection.ID: [Raindrop.ID: NSMenuItem]] { get set }
 		var emptyItems: [Screen.ItemID: NSMenuItem] { get set }
 		var loadingItems: [Screen.ItemID: NSMenuItem] { get set }
 	}
 }
 
 public extension RaindropList.View {
-	var raindropItems: [Raindrop.ID : NSMenuItem] {
+	var raindropItems: [Collection.ID: [Raindrop.ID : NSMenuItem]] {
 		get { [:] }
 		set {}
 	}
@@ -47,26 +47,28 @@ public extension RaindropList.View {
 }
 
 public extension RaindropList.View where Screen.ItemID == Collection.ID {
-	func makeMenuItem(for raindrop: Raindrop, with screen: Screen) -> NSMenuItem {
+	func makeMenuItem(for raindrop: Raindrop, in collection: Collection, with screen: Screen) -> NSMenuItem {
 		let item = NSMenuItem()
 		item.image = screen.icon(for: raindrop)
 		item.target = self
 		item.action = raindropAction
 		item.representedObject = raindrop
-		raindropItems[raindrop.id] = item
+		raindropItems[collection.id, default: [:]][raindrop.id] = item
 		return item
 	}
 
 	func raindropItems(for collection: Collection, with screen: Screen) -> [NSMenuItem] {
 		if collection.loadedRaindrops.isEmpty {
-			if screen.isUpdatingRaindrops(collection.id)  {
+			if screen.isLoadingRaindrops(collection.id)  {
 				[loadingItem(for: collection.id, with: screen)]
 			} else {
 				[emptyItem(for: collection.id, with: screen)]
 			}
 		} else {
 			collection.loadedRaindrops.map { raindrop in
-				let item = raindropItems[raindrop.id] ?? makeMenuItem(for: raindrop, with: screen)
+				let item =
+					raindropItems[collection.id]?[raindrop.id] ??
+					makeMenuItem(for: raindrop, in: collection, with: screen)
 				item.title = raindrop.title
 				return item
 			}
