@@ -1,37 +1,35 @@
 // Copyright © Fleuronic LLC. All rights reserved.
 
+import InitMacro
+import Workflow
+
 import struct Raindrop.Raindrop
 import struct Raindrop.Tag
 import struct RaindropAPI.API
 import struct RaindropService.TagWorker
 import struct RaindropService.RaindropWorker
-import class Workflow.RenderContext
-import protocol Workflow.Workflow
-import protocol Workflow.WorkflowAction
 import protocol RaindropService.TagSpec
 import protocol RaindropService.RaindropSpec
 
-public extension TagList {
-	struct Workflow<Service: TagSpec & RaindropSpec> where 
-		Service.TagLoadingResult == Tag.LoadingResult,
-		Service.RaindropLoadingResult == Raindrop.LoadingResult {
+extension TagList {
+	@Init public struct Workflow<Service: TagSpec & RaindropSpec> {
 		private let service: Service
-		
-		public init(service: Service) {
-			self.service = service
-		}
 	}
 }
 
 // MARK: -
 extension TagList.Workflow {
 	enum Action: Equatable {
-		case updateTags
-		case showTags([Tag])
-		case updateRaindrops(tagName: String, count: Int)
-		case showRaindrops([Raindrop], tagName: String)
-		case logTagError(Tag.LoadingResult.Error)
-		case logRaindropError(Raindrop.LoadingResult.Error)
+		case loadTags
+		case updateTags([Tag])
+		case finishLoadingTags
+		case handleTagLoadingError(Service.TagLoadingResult.Failure)
+
+		case loadRaindrops(tagName: String, count: Int)
+		case updateRaindrops([Raindrop], tagName: String)
+		case finishLoadingRaindrops(tagName: String)
+		case handleRaindropLoadingError(Service.RaindropLoadingResult.Failure)
+
 		case openURL(Raindrop)
 	}
 }
@@ -42,14 +40,14 @@ extension TagList.Workflow: Workflow {
 
 	public struct State {
 		var tags: [Tag]
-		var isUpdatingTags: Bool
+		var isLoadingTags: Bool
 		var updatingTags: [String: Int]
 	}
 
 	public func makeInitialState() -> State {
 		.init(
 			tags: [],
-			isUpdatingTags: false,
+			isLoadingTags: true,
 			updatingTags: [:]
 		)
 	}
@@ -59,7 +57,7 @@ extension TagList.Workflow: Workflow {
 		context: RenderContext<Self>
 	) -> TagList.Screen {
 		context.render(
-			workflows: state.isUpdatingTags ? [tagWorker.asAnyWorkflow()] : [],
+			workflows: state.isLoadingTags ? [tagWorker.asAnyWorkflow()] : [],
 			keyedWorkflows: .init(
 				uniqueKeysWithValues: state.updatingTags.map { name, count in
 					(name, raindropWorker(forTagNamed: name, count: count).asAnyWorkflow())
@@ -67,12 +65,14 @@ extension TagList.Workflow: Workflow {
 			)
 		) { sink in
 			.init(
-				updateRaindrops: { sink.send(.updateRaindrops(tagName: $0, count: $1)) },
-				isUpdatingRaindrops: state.updatingTags.keys.contains,
+				loadRaindrops: { sink.send(.loadRaindrops(tagName: $0, count: $1)) },
+				isLoadingRaindrops: state.updatingTags.keys.contains,
+				finishLoadingRaindrops: { sink.send(.finishLoadingRaindrops(tagName: $0)) },
 				selectRaindrop: { sink.send(.openURL($0)) },
 				tags: state.tags,
-				updateTags: { sink.send(.updateTags) },
-				isUpdatingTags: state.isUpdatingTags
+				loadTags: { sink.send(.loadTags) },
+				isLoadingTags: state.isLoadingTags,
+				finishLoadingTags: { sink.send(.finishLoadingTags) }
 			)
 		}
 	}
@@ -83,8 +83,9 @@ private extension TagList.Workflow {
 	var tagWorker: TagWorker<Service, Action> {
 		.init(
 			service: service,
-			success: Action.showTags,
-			failure: Action.logTagError
+			success: { .updateTags($0) },
+			failure: { .handleTagLoadingError($0) },
+			completion: .finishLoadingTags
 		)
 	}
 
@@ -93,8 +94,9 @@ private extension TagList.Workflow {
 			service: service,
 			source: .tag(name: name),
 			count: count,
-			success: { Action.showRaindrops($0, tagName: name) },
-			failure: Action.logTagError
+			success: { .updateRaindrops($0, tagName: name) },
+			failure: { .handleRaindropLoadingError($0) },
+			completion: .finishLoadingRaindrops(tagName: name)
 		)
 	}
 }
@@ -118,19 +120,21 @@ extension TagList.Workflow.Action: WorkflowAction {
 
 	func apply(toState state: inout WorkflowType.State) -> Raindrop? {
 		switch self {
-		case .updateTags:
-			state.isUpdatingTags = true
-		case let .showTags(tags):
+		case .loadTags:
+			state.isLoadingTags = true
+		case let .updateTags(tags):
 			state.tags = tags
-			state.isUpdatingTags = false
-		case let .updateRaindrops(tagName, tagCount):
+		case .finishLoadingTags:
+			state.isLoadingTags = false
+		case let .loadRaindrops(tagName, tagCount):
 			state.updatingTags[tagName] = tagCount
-		case let .showRaindrops(raindrops, tagName):
+		case let .updateRaindrops(raindrops, tagName):
 			state.update(with: raindrops, taggedWithTagNamed: tagName)
+		case let .finishLoadingRaindrops(tagName: tagName):
 			state.updatingTags.removeValue(forKey: tagName)
-		case let .logTagError(error):
+		case let .handleTagLoadingError(error):
 			print(error)
-		case let .logRaindropError(error):
+		case let .handleRaindropLoadingError(error):
 			print(error)
 		case let .openURL(raindrop):
 			return raindrop
